+++ tests/endpoints/update-vested-tokens.ts
@@ -53,10 +53,8 @@ export function test() {
       expect(
         vestingInfoAfter.cumulativeVestedAmount.amount.toNumber()
       ).to.above(0);
-      expect(vestingInfoBefore.unfundedLiabilities.amount.toNumber()).to.eq(0);
-      expect(vestingInfoAfter.unfundedLiabilities.amount.toNumber()).to.above(
-        0
-      );
+      expect(vestingInfoBefore.unfundedLiability.amount.toNumber()).to.eq(0);
+      expect(vestingInfoAfter.unfundedLiability.amount.toNumber()).to.above(0);
 
       // Everything else remains with the same default values
       expect(vestingInfoAfter.totalVestingAmount.amount.toNumber()).to.eq(
@@ -72,7 +70,7 @@ export function test() {
       expect(
         vestingInfoAfter.cumulativeWithdrawnAmount.amount.toNumber()
       ).to.eq(0);
-      expect(vestingInfoAfter.vestingVaultBalance.amount.toNumber()).to.eq(0);
+      expect(vestingInfoAfter.vaultBalance.amount.toNumber()).to.eq(0);
     });
   });
 }