+++ programs/vesting-treasury/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2022-07-29
+
+### Added
+
+- Endpoint `fund_vesting_vault`
+
 ## [0.3.0] - 2022-07-27
 
 ### Added
@@ -13,6 +19,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 - Endpoint `update_vested_tokens`
 - Method `update_vested_tokens` in struct `Vesting` which contains
   core logic for the endpoint
+- Method `update_unfunded_liability` in struct `Vesting`
 - Chrono external crate to handle datetime calculations
 
 ## [0.2.0] - 2022-07-27